+++ packages/orq-rc/src/models/components/partdoneevent.ts
@@ -79,7 +79,7 @@ export const PartReasoningPart$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("reasoning_01kf08b16xpa9m9d82k216qs3d"),
+  _id: z.string().default("reasoning_01kf0t8nq5kw2dx7v545x2qfjb"),
   metadata: z.record(z.any()).optional(),
   kind: PartKind$inboundSchema,
   reasoning: z.string(),
@@ -104,7 +104,7 @@ export const PartReasoningPart$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   PartReasoningPart
 > = z.object({
-  id: z.string().default("reasoning_01kf08b16xpa9m9d82k216qs3d"),
+  id: z.string().default("reasoning_01kf0t8nq5kw2dx7v545x2qfjb"),
   metadata: z.record(z.any()).optional(),
   kind: PartKind$outboundSchema,
   reasoning: z.string(),