+++ packages/orq-rc/src/models/operations/listidentities.ts
@@ -310,7 +310,7 @@ export const ListIdentitiesData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-01-15T07:18:41.663Z",
+    "2026-01-15T12:31:58.759Z",
   ).transform(v => new Date(v)),
   metrics: z.lazy(() => ListIdentitiesMetrics$inboundSchema),
 }).transform((v) => {
@@ -349,7 +349,7 @@ export const ListIdentitiesData$outboundSchema: z.ZodType<
   tags: z.array(z.string()).optional(),
   metadata: z.record(z.any()).optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2026-01-15T07:18:41.663Z"))
+  updated: z.date().default(() => new Date("2026-01-15T12:31:58.759Z"))
     .transform(v => v.toISOString()),
   metrics: z.lazy(() => ListIdentitiesMetrics$outboundSchema),
 }).transform((v) => {